Sysl generate sequence diagram
* Add test case
* Fix code comments
* Decouple codes of sequence diagram
