+++ sysl2/sysl/seqs_test.go
@@ -139,104 +139,6 @@ func TestLoadApp(t *testing.T) {
 	assert.Equal(t, []string{"user_id"}, expectedParams)
 }
 
-func TestTransformBlackBoxes(t *testing.T) {
-	type args struct {
-		blackboxes []*sysl.Attribute
-	}
-
-	eltFirst := []*sysl.Attribute{
-		{
-			Attribute: &sysl.Attribute_S{
-				S: "Value A",
-			},
-		},
-		{
-			Attribute: &sysl.Attribute_S{
-				S: "Value B",
-			},
-		},
-	}
-	attrFirst := &sysl.Attribute{
-		Attribute: &sysl.Attribute_A{
-			A: &sysl.Attribute_Array{
-				Elt: eltFirst,
-			},
-		},
-	}
-	eltSecond := []*sysl.Attribute{
-		{
-			Attribute: &sysl.Attribute_S{
-				S: "Value C",
-			},
-		},
-		{
-			Attribute: &sysl.Attribute_S{
-				S: "Value D",
-			},
-		},
-	}
-	attrSecond := &sysl.Attribute{
-		Attribute: &sysl.Attribute_A{
-			A: &sysl.Attribute_Array{
-				Elt: eltSecond,
-			},
-		},
-	}
-
-	tests := []struct {
-		name string
-		args args
-		want [][]string
-	}{
-		{
-			name: "Case-Null",
-			args: args{blackboxes: []*sysl.Attribute{}},
-			want: [][]string{},
-		},
-		{
-			name: "Case-ConvertSuccess",
-			args: args{blackboxes: []*sysl.Attribute{attrFirst, attrSecond}},
-			want: [][]string{{"Value A", "Value B"}, {"Value C", "Value D"}},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := transformBlackBoxes(tt.args.blackboxes); !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("transformBlackBoxes() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func TestParseBlackBoxesFromArgument(t *testing.T) {
-	type args struct {
-		blackboxFlags []string
-	}
-	tests := []struct {
-		name string
-		args args
-		want [][]string
-	}{
-		{
-			name: "Case-Null",
-			args: args{[]string{}},
-			want: [][]string{},
-		},
-		{
-			name: "Case-ConvertSuccess",
-			args: args{[]string{"Value A,Value B", "Value C,Value D"}},
-			want: [][]string{{"Value A", "Value B"}, {"Value C", "Value D"}},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := parseBlackBoxesFromArgument(tt.args.blackboxFlags); !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("parseBlackBoxesFromArgument() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
 func TestSimpleParserFmtEp(t *testing.T) {
 	type fields struct {
 		self string
@@ -396,93 +298,6 @@ func TestSimpleParserFmtOutput(t *testing.T) {
 	}
 }
 
-func TestFindMatchItems(t *testing.T) {
-	type args struct {
-		origin string
-	}
-	tests := []struct {
-		name string
-		args args
-		want []string
-	}{
-		{
-			name: "Case-Null",
-			args: args{""},
-			want: nil,
-		},
-		{
-			name: "Case-Convert Success",
-			args: args{"%(appname)"},
-			want: []string{"%(appname)"},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := findMatchItems(tt.args.origin); !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("findMatchItems() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func TestRemoveWrapper(t *testing.T) {
-	type args struct {
-		origin string
-	}
-	tests := []struct {
-		name string
-		args args
-		want string
-	}{
-		{
-			name: "Case-Null",
-			args: args{""},
-			want: "",
-		},
-		{
-			name: "Case-Convert Success",
-			args: args{"%(appname)"},
-			want: "appname",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := removeWrapper(tt.args.origin); got != tt.want {
-				t.Errorf("removeWrapper() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func TestRemovePercentSymbol(t *testing.T) {
-	type args struct {
-		origin string
-	}
-	tests := []struct {
-		name string
-		args args
-		want string
-	}{
-		{
-			"Case-Null",
-			args{""},
-			"",
-		},
-		{
-			"Case-Remove Percent",
-			args{"%VariableA, %VariableB"},
-			"VariableA, VariableB",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := removePercentSymbol(tt.args.origin); got != tt.want {
-				t.Errorf("removePercentSymbol() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
 func Test_constructSimpleParser(t *testing.T) {
 	type args struct {
 		former string
@@ -513,70 +328,6 @@ func Test_constructSimpleParser(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			if got := constructSimpleParser(tt.args.former, tt.args.latter); !reflect.DeepEqual(got, tt.want) {
 				t.Errorf("constructSimpleParser() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func Test_mergeAttributes(t *testing.T) {
-	type args struct {
-		app   map[string]*sysl.Attribute
-		edpnt map[string]*sysl.Attribute
-	}
-
-	appAttr := &sysl.Attribute{
-		Attribute: &sysl.Attribute_S{
-			S: "Value A",
-		},
-	}
-	appMap := map[string]*sysl.Attribute{
-		"app": appAttr,
-	}
-	epAttr := &sysl.Attribute{
-		Attribute: &sysl.Attribute_S{
-			S: "Value B",
-		},
-	}
-	epMap := map[string]*sysl.Attribute{
-		"ep": epAttr,
-	}
-	tests := []struct {
-		name string
-		args args
-		want map[string]*sysl.Attribute
-	}{
-		{
-			"Case-Null",
-			args{},
-			map[string]*sysl.Attribute{},
-		},
-		{
-			"Case-Merge app",
-			args{appMap, map[string]*sysl.Attribute{}},
-			map[string]*sysl.Attribute{
-				"app": appAttr,
-			},
-		},
-		{
-			"Case-Merge ep",
-			args{map[string]*sysl.Attribute{}, epMap},
-			map[string]*sysl.Attribute{
-				"ep": epAttr,
-			},
-		},
-		{
-			"Case-Merge app and ep",
-			args{appMap, epMap},
-			map[string]*sysl.Attribute{
-				"app": appAttr,
-				"ep": epAttr,
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := mergeAttributes(tt.args.app, tt.args.edpnt); !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("mergeAttributes() = %v, want %v", got, tt.want)
 			}
 		})
 	}
